strands-agents/requirements.txt
streamlit==1.31.0
streamlit-oauth==0.1.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
requests==2.31.0
python-dotenv==1.0.0
boto3==1.34.0
mem0ai==0.1.0
strands-agents==0.1.0
qdrant-client==1.7.0
